dm_model/requirements.txt
torch==1.13.1
torchvision
pillow
numpy
